backend/fisio_find/gestion_usuarios/urls.py
from django.urls import path
from .views import PatientRegisterView
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    path('api/v1/patient/register', PatientRegisterView.as_view(), name='patient_register'),
    path('api/v1/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/v1/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]


