backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
djangorestframework_simplejwt==5.4.0
django-environ==0.12.0
django-cors-headers==4.7.0
python-dotenv==1.0.1
pillow==11.1.0